Add doc tests to handler setters
